Introduced BC. AlBlockManager's getHtml method has been declared final. To change the default rendered content you must now override the new protected method renderHtml which works exactly as the getHtml method worked before this change.
Internal bundles have been updated to reflect this change.

Introduced the new property editorDisabled and its getter and setter to control the editor rendering for the
AlBlockManager object. By default the editor is rendered. In this way it is possible to avoid rendering the
blocks' editor required attributes when the site is deployed.

Updated the SlotRendererExtension and the templates to reflect these changes.