Fix collision issue in al-inline-list.js
Changed the collision parameters from 'fit fit' to 'none none' to fix issue of add and delete button containers not being positioned correctly for list elements that are not in the browser window. The 'fit' option causes the button container to be placed at the bottom of the browser window for all list elements that are not within the viewable browser window. The offscreen element's add/delete button containers are usually 'stacked' behind each other as they all end up with the same y position.